test-requirements.txt
ray
py-cpuinfo
pandas
numpy
pytest
nose